Refactored authentication scheme
