--- front/src/stores/user.js
@@ -1,10 +1,0 @@
-import { defineStore } from 'pinia'
-import { ref } from 'vue'
-
-export const useUserStore = defineStore('user', () => {
-    const tokenData = ref(null);
-
-    return {
-        tokenData,
-    }
-})
